% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Assess whether a gene has allele specific translation in a given individual.}
\usage{
analyze(file.name, gene.name, expected.value = NULL, grouped = FALSE,
  grouped.fractions = NULL, zoom.range = NULL, nsims = 1000,
  output.pvals = "output_pvals.csv", output.plots = "output_plots.pdf")
}
\arguments{
\item{file.name}{Filename of the ddPCR counts data file to be read in. The
data must be formatted in the following manner: each row corresponds to a
ddPCR well, and there must be exactly 7 columns - well ID, fraction ID,
Ch1+/Ch2+ counts, Ch1+/Ch2- counts, Ch1-/Ch2+ counts, Ch1-/Ch2- counts, and
proportion of total concentration.}

\item{gene.name}{A string indicating the name of gene being studied.}

\item{expected.value}{A number indicating the overall Ch1+/Ch2+ ratio that
would be expected if there were no allele specific translation. The user
can include if this value is already known. If \code{NULL}, the function will
automatically calculate an expected value by summing over all fractions.
Defaults to \code{NULL}.}

\item{grouped}{A boolean indicating whether to group fractions by number of
polysomes. Defaults to \code{FALSE}.}

\item{grouped.fractions}{A vector indicating the number of experimental
fractions corresponding to each ribosomal fraction. Defaults to \code{NULL}.}

\item{zoom.range}{A vector of 2 numbers that indicate the particular range of
experimental fractions to graph. Defaults to \code{NULL}, which corresponds to
plot all graphs. NOTE: zoom.range can only be used with fractions that are
NOT grouped by ribosomes.}

\item{nsims}{The number of bootstrap simulations to run for each fraction}

\item{output.pvals}{File name of the csv output file}

\item{output.plots}{File name of the pdf output file}
}
\value{
There are two outputs. The first is a CSV file with each fraction /
  group, it's p-value, and its proportion of the total concentration. The
  second output is a PDF containing two aligned graphs. One graph corresponds
  to a boxplot of all bootstrapped Ch1+/Ch2+ ratios relative to the expected
  value (either calculated or inputted by the users). This makes it visually
  clear which fractions are significantly different (p < 0.05 for a
  two-tailed test) from the expected ratios. The other graph is a barplot
  that graphs the ddPCR concentrations from each fraction so that users can
  see the distribution and relative concentrations of RNA across all
  fractions (from the early fractions with no ribosomes to the heavier
  fractions with 5+ ribosomes).
}
\description{
This package helps the user assess whether a gene in a particular
heterozygous individual has allele-specific translation based on data from an
assay that combines polysome profiling and digital droplet PCR (ddPCR). It
uses bootstrapping techniques to determine whether the amount of mutant and
wildtype transcripts in each fraction (directly from the polysome profiling
or grouped by number of ribosomes) is significantly different from the
expected values. Each fraction corresponds to a particular weight (heavier
transcripts contain a greater number of ribosomes, indicating greater
translation).
}
\examples{
  analyze (filename = "gene1_ddPCR_data.csv", gene.name = "GENE1")
  analyze (filename = "gene2_ddPCR_data.csv", gene.name = "GENE2", zoom.range = c(5, 16))
  analyze (filename = "gene3_ddPCR_data.csv", gene.name = "GENE3", expected.value = 1.2, grouped = TRUE, grouped.fractions = c(7, 3, 1, 1, 1, 5))
}

