% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_candidates.R
\name{find_candidates}
\alias{find_candidates}
\title{Determining which SNPs are candidates for allele specific translation}
\usage{
find_candidates(RNA.file, RIBO.file, output.csv, output.pdf, alpha = 0.05,
  pcounts = 50, c.threshold = 0.9, d.threshold = 0.1, num.sims = 1000)
}
\arguments{
\item{RNA.file}{Filename of the RNA-Seq counts for each SNP. Each row is either the maternal
or paternal allele of a SNP and its associated number of RNA-Seq reads at each position.
The number of columns is defined by the size of the window (i.e. for RNA, usually around
75 nucleotides.}

\item{RIBO.file}{Filename of the Ribosome profiling counts for each SNP. See \code{RNA.file}
for more details.}

\item{output.csv}{Filename of the output csv file where the numerical output (effect sizes
and p values) will be stored.}

\item{output.pdf}{Filename of the output pdf file where the graphical output (histograms
of the RNA-Seq and ribosome profiling bootstrapped ratio distributions) will be
stored}

\item{alpha}{Threshold for what p-values are considered significant. Default is 0.05.}

\item{pcounts}{Pseudocounts for regularization when calculating the maternal / total ratio
counts (for each bootstrapped sample) so tha the RNA-Seq ratio and ribosomal
profiling ratio can be comparable. Default value is 50. Should be a value between
~30 (the window size ribosome profiling) and ~75 (the window size for RNA-seq).}

\item{c.threshold}{Threshold for what Cliff's Delta values are considered significant.
Default is 0.9.}

\item{d.threshold}{Threshold for what secondary effect size difference is significant.
Default is 0.1 (i.e. a minimum of a 10% difference between the allelic ratio
of the RNA-seq counts and ribosome profiling counts).}

\item{num.sims}{Number of times to bootstrap. Default is 1000.}
}
\value{
Two files. 1) A .csv file with all candidate SNPs and their associated
       p-values, Cliff's Delta, and secondary effect size values (9 columns total).
       2) A .pdf file with a histogram of the bootstrapped ratio distributions
       for RNA-Seq (red) and ribosome profiling (blue) for each candidate SNP. This makes
       it possible to easily visualize how much overlap there are between the
       two distributions.
}
\description{
This function helps users identify which genetic variants are candidates for
allele-specific translation (i.e. may lead to differential translation) by
using RNA-Seq and Ribosome profiling data from all heterozygous SNPs in the
transcriptome of a given individual.
}
\examples{
find_candidates("RNA.all.csv", "RIBO.all.csv", "final.effsize.csv", "final.plots.pdf")
find_candidates("RNA.all.csv", "RIBO.all.csv", "final.effsize.csv", "final.plots.pdf", num.sims = 5000, c.threshold = 0.5)

}

